Add usage indicator to run script (#2)
Although the anagram class provides an argument length check and emits
usage information, this fails when invoking with "$1" "$2", as two empty
words are provided as the params. This makes `args.length == 2` but
still fails when the application attempts to load the dictionary file
named: ''.

+++ debug_assignment/run.sh
@@ -1,4 +1,11 @@
 #!/usr/bin/env bash
+if [ $# -ne 2 ]; then
+  echo "Usage:"
+  echo "$0 <phrase> <dictionaryfile>"
+  exit -1
+fi;
+
+
 echo $1 $2
 
 rm -fr classes
